+++ lib/avx10_t1/mb_mgr_avx10_t1.c
@@ -256,6 +256,9 @@
 #define SUBMIT_JOB_ZUC_NEA6 submit_job_zuc_nea6_gfni_sse
 #define FLUSH_JOB_ZUC_NEA6  flush_job_zuc_nea6_gfni_sse
 
+/* SNOW5G */
+#define SUBMIT_JOB_SNOW5G snow_5g_sse
+
 /* SNOW3G UEA2 & UIA2 */
 static IMB_JOB *
 submit_snow3g_uea2_job_vaes_avx512(IMB_MGR *state, IMB_JOB *job)