+++ ReleaseNotes.md
@@ -23,6 +23,7 @@ Library
 - Removed KASUMI bit-length support and corresponding API.
 - Added support for zero-length messages in HMAC-SHA and HMAC-MD5 algorithms.
 - Added ZUC-NEA6 SSE implementation.
+- Added SNOW5G-NEA4 SSE type 1 implementation.
 
 Test Applications
 - Added `--offset` parameter to `imb-xvalid` application.